Small fixes of type comparison

+++ DoubleColumn.cpp
@@ -10,7 +10,7 @@ std::string DoubleColumn::getType() const {
 }
 
 std::string DoubleColumn::at(const unsigned int& index) const {
-    if(index >= (int)data.size()) {
+    if(index >= data.size()) {
         return ""; //todo
     }
 
@@ -55,7 +55,7 @@ DoubleColumn::getRowsIndicesWith(const std::string& value) const {
 }
 
 void DoubleColumn::updateRowByIndex(const unsigned int& index, const std::string& value) {
-    if(index >= (int)this->data.size()) {
+    if(index >= this->data.size()) {
         return; //out of range
     }
 
@@ -70,7 +70,7 @@ void DoubleColumn::updateRowByIndex(const unsigned int& index, const std::string
 }
 
 void DoubleColumn::deleteRowByIndex(const unsigned int& index) {
-    int n = (int)this->data.size();
+    unsigned int n = this->data.size();
 
     if(index >= n) {
         return; //out of range
